# bounties_api/static/autogenerated_swagger.yml
swagger: '2.0'
info:
  title: Bounties Network API
  description: Bounties Network API documentation
  termsOfService: https://explorer.bounties.network/tos
  contact:
    email: contact@bounties.network
  license:
    name: MIT License
  version: v1
host: 127.0.0.1:8000
schemes:
  - http
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
security:
  - Basic: []
paths:
...
  /user/languages/:
    get:
      operationId: user_languages_list
      description: Returns ordered languages
      parameters:
        - name: limit
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
        - name: offset
          in: query
          description: The initial index from which to return the results.
          required: false
          type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
              - count
              - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
              previous:
                type: string
                format: uri
              results:
                type: array
                items:
                  $ref: '#/definitions/Language'
      tags:
        - user
    parameters: []
...
  /user/skills/:
    get:
      operationId: user_skills_list
      description: Returns ordered skills
      parameters:
        - name: ordering
          in: query
          description: Which field to use when ordering the results.
          required: false
          type: string
        - name: search
          in: query
          description: A search term.
          required: false
          type: string
        - name: limit
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
        - name: offset
          in: query
          description: The initial index from which to return the results.
          required: false
          type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
              - count
              - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
              previous:
                type: string
                format: uri
              results:
                type: array
                items:
                  $ref: '#/definitions/RankedSkill'
      tags:
        - user
    parameters: []
...
definitions:
...
  Language:
    required:
      - name
      - normalized_name
    type: object
    properties:
      name:
        title: Name
        type: string
        maxLength: 128
        minLength: 1
      normalized_name:
        title: Normalized name
        type: string
        maxLength: 128
        minLength: 1
  RankedSkill:
    required:
      - name
      - normalized_name
      - total_count
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        maxLength: 128
        minLength: 1
      normalized_name:
        title: Normalized name
        type: string
        maxLength: 128
        minLength: 1
      total_count:
        title: Total count
        type: integer
        maximum: 2147483647
        minimum: -2147483648